Trait smoldot_light::platform::Platform

pub trait Platform: Send + 'static {
    type Delay: Future<Output = ()> + Unpin + Send + 'static;
    type Yield: Future<Output = ()> + Unpin + Send + 'static;
    type Instant: Clone + Add<Duration, Output = Self::Instant> + Sub<Self::Instant, Output = Duration> + PartialOrd + Ord + PartialEq + Eq + Send + Sync + 'static;
    type Connection: Send + Sync + 'static;
    type Stream: Send + 'static;
    type ConnectFuture: Future<Output = Result<PlatformConnection<Self::Stream, Self::Connection>, ConnectError>> + Unpin + Send + 'static;
    type StreamDataFuture<'a>: Future<Output = ()> + Unpin + Send + 'a;
    type NextSubstreamFuture<'a>: Future<Output = Option<(Self::Stream, PlatformSubstreamDirection)>> + Unpin + Send + 'a;

    fn now_from_unix_epoch() -> Duration;
    fn now() -> Self::Instant;
    fn sleep(duration: Duration) -> Self::Delay;
    fn sleep_until(when: Self::Instant) -> Self::Delay;
    fn yield_after_cpu_intensive() -> Self::Yield;
    fn connect(url: &str) -> Self::ConnectFuture;
    fn open_out_substream(connection: &mut Self::Connection);
    fn next_substream(
        connection: &mut Self::Connection
    ) -> Self::NextSubstreamFuture<'_>;
    fn wait_more_data(stream: &mut Self::Stream) -> Self::StreamDataFuture<'_>;
    fn read_buffer(stream: &mut Self::Stream) -> ReadBuffer<'_>;
    fn advance_read_cursor(stream: &mut Self::Stream, bytes: usize);
    fn send(stream: &mut Self::Stream, data: &[u8]);
}

Access to a platform’s capabilities.

Required Associated Types

type Delay: Future<Output = ()> + Unpin + Send + 'static

type Yield: Future<Output = ()> + Unpin + Send + 'static

type Instant: Clone + Add<Duration, Output = Self::Instant> + Sub<Self::Instant, Output = Duration> + PartialOrd + Ord + PartialEq + Eq + Send + Sync + 'static

type Connection: Send + Sync + 'static

A multi-stream connection.

This object is merely a handle. The underlying connection should be dropped only after the Connection and all its associated substream objects (Platform::Stream) have been dropped.

type Stream: Send + 'static

type ConnectFuture: Future<Output = Result<PlatformConnection<Self::Stream, Self::Connection>, ConnectError>> + Unpin + Send + 'static

type StreamDataFuture<'a>: Future<Output = ()> + Unpin + Send + 'a

type NextSubstreamFuture<'a>: Future<Output = Option<(Self::Stream, PlatformSubstreamDirection)>> + Unpin + Send + 'a

Required Methods

fn now_from_unix_epoch() -> Duration

Returns the time elapsed since the Unix Epoch (i.e. 00:00:00 UTC on 1 January 1970), ignoring leap seconds.

Panic

Panics if the system time is configured to be below the UNIX epoch. This situation is a very very niche edge case that isn’t worth handling.

fn now() -> Self::Instant

Returns an object that represents “now”.

fn sleep(duration: Duration) -> Self::Delay

Creates a future that becomes ready after at least the given duration has elapsed.

fn sleep_until(when: Self::Instant) -> Self::Delay

Creates a future that becomes ready after the given instant has been reached.

fn yield_after_cpu_intensive() -> Self::Yield

Should be called after a CPU-intensive operation in order to yield back control.

This function can be implemented as no-op on platforms where this is irrelevant.

fn connect(url: &str) -> Self::ConnectFuture

Starts a connection attempt to the given multiaddress.

The multiaddress is passed as a string. If the string can’t be parsed, an error should be returned where ConnectError::is_bad_addr is true.

fn open_out_substream(connection: &mut Self::Connection)

Queues the opening of an additional outbound substream.

The substream, once opened, must be yielded by Platform::next_substream.

Calls to this function should be ignored if the connection has already been killed by the remote.

Note: No mechanism exists in this API to handle the situation where a substream fails to open, as this is not supposed to happen. If you need to handle such a situation, either try again opening a substream again or reset the entire connection.

fn next_substream(
connection: &mut Self::Connection
) -> Self::NextSubstreamFuture<'_>

Waits until a new incoming substream arrives on the connection.

This returns both inbound and outbound substreams. Outbound substreams should only be yielded once for every call to Platform::open_out_substream.

The future can also return None if the connection has been killed by the remote. If the future returns None, the user of the Platform should drop the Connection and all its associated Streams as soon as possible.

fn wait_more_data(stream: &mut Self::Stream) -> Self::StreamDataFuture<'_>

Returns a future that becomes ready when either the read buffer of the given stream contains data, or the remote has closed their sending side.

The future is immediately ready if data is already available or the remote has already closed their sending side.

fn read_buffer(stream: &mut Self::Stream) -> ReadBuffer<'_>

Gives access to the content of the read buffer of the given stream.

fn advance_read_cursor(stream: &mut Self::Stream, bytes: usize)

Discards the first bytes bytes of the read buffer of this stream. This makes it possible for the remote to send more data.

Panic

Panics if there aren’t enough bytes to discard in the buffer.

fn send(stream: &mut Self::Stream, data: &[u8])

Queues the given bytes to be sent out on the given connection.

Note: In the case of PlatformConnection::MultiStreamWebRtc, be aware that there exists a limit to the amount of data to send at once. The data parameter is guaranteed to fit within that limit. Due to the existence of this limit, the implementation of this function shouldn’t attempt to save function calls by performing internal buffering and batching multiple calls into one.