smoldot_light::platform

Module with_buffers

Augments an implementation of AsyncRead and AsyncWrite with a read buffer and a write buffer.

While this module is generic, the targeted use-case is TCP connections.

Structs

ReadWriteAccess

See WithBuffers::read_write_access.

WithBuffers

Holds an implementation of AsyncRead and AsyncWrite, alongside with a read buffer and a write buffer.