Struct smoldot_light::Client

pub struct Client<TPlat: Platform, TChain = ()> { /* private fields */ }

Holds a list of chains, connections, and JSON-RPC services.

Implementations

impl<TPlat: Platform, TChain> Client<TPlat, TChain>

pub fn new(config: ClientConfig) -> Self

Initializes the smoldot client.

pub fn add_chain(
&mut self,
config: AddChainConfig<'_, TChain, impl Iterator<Item = ChainId>>
) -> Result<AddChainSuccess, String>

Adds a new chain to the list of chains smoldot tries to synchronize.

pub fn remove_chain(&mut self, id: ChainId) -> TChain

Removes the chain from smoldot. This instantaneously and silently cancels all on-going JSON-RPC requests and subscriptions.

The provided ChainId is now considered dead. Be aware that this same ChainId might later be reused if Client::add_chain is called again.

While from the API perspective it will look like the chain no longer exists, calling this function will not actually immediately disconnect from the given chain if it is still used as the relay chain of a parachain.

If the JsonRpcResponses object that was returned when adding the chain is still alive, JsonRpcResponses::next will now return None.

pub fn chain_user_data_mut(&mut self, chain_id: ChainId) -> &mut TChain

Returns the user data associated to the given chain.

Panic

Panics if the ChainId is invalid.

pub fn json_rpc_request(
&mut self,
json_rpc_request: impl Into<String>,
chain_id: ChainId
) -> Result<(), HandleRpcError>

Enqueues a JSON-RPC request towards the given chain.

Since most JSON-RPC requests can only be answered asynchronously, the request is only queued and will be decoded and processed later.

Returns an error if the node is overloaded and is capable of processing more JSON-RPC requests before some time has passed or the AddChainSuccess::json_rpc_responses stream is emptied.

Also returns an error if the request could not be parsed as a valid JSON-RPC request, as in that situation smoldot is unable to send back a corresponding JSON-RPC error message.

Panic

Panics if the ChainId is invalid, or if AddChainConfig::disable_json_rpc was true when adding the chain.