Struct smoldot_light::PeerId

pub struct PeerId { /* private fields */ }

Identifier of a node of the network.

The data is a multihash of the public key of the peer.

Implementations

impl PeerId

pub fn from_public_key(key: &PublicKey) -> PeerId

Builds the PeerId corresponding to a public key.

pub fn from_bytes(
data: Vec<u8, Global>
) -> Result<PeerId, (FromBytesError, Vec<u8, Global>)>

Checks whether data is a valid PeerId.

In case of error, returns the bytes passed as parameter in addition to the error.

pub fn into_bytes(self) -> Vec<u8, Global>

Returns a raw bytes representation of this PeerId.

pub fn as_bytes(&self) -> &[u8]

Returns a raw bytes representation of this PeerId.

pub fn to_base58(&self) -> String

Returns a base-58 encoded string of this PeerId.