smoldot_light
Module platform

Re-exports

pub use default::DefaultPlatform;

Modules

address_parse

default

Implementation of the PlatformRef trait that leverages the operating system.

read_write

with_buffers

Augments an implementation of AsyncRead and AsyncWrite with a read buffer and a write buffer.

Structs

MultiStreamWebRtcConnection

Established multistream connection information. See PlatformRef::connect_multistream.

WithPrefix

Implementation of a PlatformRef that wraps around another platform and adds a prefix before every log line and task name.

Enums

Address

Address passed to PlatformRef::connect_stream.

ConnectionType

Connection type passed to PlatformRef::supports_connection_type.

LogLevel

Log level of a log entry.

MultiStreamAddress

Address passed to PlatformRef::connect_multistream.

SubstreamDirection

Direction in which a substream has been opened. See PlatformRef::next_substream.

Traits

PlatformRef

Access to a platform’s capabilities.