smoldot_light::platform

Struct WithPrefix

pub struct WithPrefix<T> { /* private fields */ }

Implementation of a PlatformRef that wraps around another platform and adds a prefix before every log line and task name.

Implementations

impl<T> WithPrefix<T>

pub const fn new(prefix: String, inner: T) -> Self

Builds a new WithPrefix.