smoldot_light

Struct Client

pub struct Client<TPlat: PlatformRef, TChain = ()> { /* private fields */ }

Holds a list of chains, connections, and JSON-RPC services.

Implementations

impl<TPlat: PlatformRef, TChain> Client<TPlat, TChain>

pub const fn new(platform: TPlat) -> Self

Initializes the smoldot client.

pub fn add_chain( &mut self, config: AddChainConfig<'_, TChain, impl Iterator<Item = ChainId>>, ) -> Result<AddChainSuccess<TPlat>, AddChainError>

Adds a new chain to the list of chains smoldot tries to synchronize.

Returns an error in case something is wrong with the configuration.

pub fn remove_chain(&mut self, id: ChainId) -> TChain

Removes the chain from smoldot. This instantaneously and silently cancels all on-going JSON-RPC requests and subscriptions.

The provided ChainId is now considered dead. Be aware that this same ChainId might later be reused if Client::add_chain is called again.

While from the API perspective it will look like the chain no longer exists, calling this function will not actually immediately disconnect from the given chain if it is still used as the relay chain of a parachain.

If the JsonRpcResponses object that was returned when adding the chain is still alive, JsonRpcResponses::next will now return None.

pub fn json_rpc_request( &mut self, json_rpc_request: impl Into<String>, chain_id: ChainId, ) -> Result<(), HandleRpcError>

Enqueues a JSON-RPC request towards the given chain.

Since most JSON-RPC requests can only be answered asynchronously, the request is only queued and will be decoded and processed later.

Returns an error if the number of requests that have been sent but whose answer hasn’t been pulled with JsonRpcResponses::next is superior or equal to the value that was passed through AddChainConfigJsonRpc::Enabled::max_pending_requests. In that situation, the API user is encouraged to stop sending requests and start pulling answers with JsonRpcResponses::next.

Passing u32::MAX to AddChainConfigJsonRpc::Enabled::max_pending_requests is a good way to avoid errors here, but this should only be done if the JSON-RPC client is trusted.

If the JSON-RPC request is not a valid JSON-RPC request, a JSON-RPC error response with an id equal to null is later generated, in accordance with the JSON-RPC specification.

Panic

Panics if the ChainId is invalid, or if AddChainConfig::json_rpc was AddChainConfigJsonRpc::Disabled when adding the chain.